AutoShutDown/requirements.txt
azureml-sdk==1.19.0
azure-mgmt-containerinstance==1.5.0